+++ R/factory.R
@@ -37,6 +37,8 @@ factory <- function(tab,
     factory_fun <- factory_DT
   } else if (settings_equal("output_factory", "dataframe")) {
     factory_fun <- factory_dataframe
+  } else if (settings_equal("output_factory", "modelsummary_markdown")) {
+    factory_fun <- factory_markdown
   }
 
   # flat header if necessary